Refactor saving and loading.
I admit, I was doing something that I shouldn't have been - loading in the constructor! I've split out loading to outside of the constructor now, and added a special hidden config option to display when Nucleus is saving or loading a particular entity. I'm hoping this either fixes, or at least makes it easier to pinpoint the loading/saving issues we're having.