Rewrite the RTP system.
There are several things going on here, but the biggest is that I've created a RTP registry module so custom RTP algorithms can be used with Nucleus.

The motivation behind this is that the RTP system for Nucleus has been a little sub par for some time. This is due to the fact that there are a lot of wants and needs for servers, but it usually boils down to wanting specific biome teleportation or some other spread of teleportation, or even location specific teleportation.

As I have no interest in doing all of this, my usual respose is to "go write your own plugin". However, the nucleus RTP handler is often still desired. So, the idea is that plugins can write their own kernels and register them, then server owners can choose the kernel they want to use in the config. Config options would be passed onto the kernel with `/rtp`.

This would expose a service to register and use kernels, as well as create a registry module for RTP.

As a side effect, this will pull out the code to pick a spot to make it more testable for robustness.

This is in order to fix #1223 by ripping the code into its own class and make it more testable, but also to enable customisation by plugin authors.

* Update QSML to 0.11.0
* Bump snapshot version to 1.5.0